Add bindings for ember destroy
With generating becoming so easy, we often seem to be using the
generator very quickly.  Sometimes we figure out a better name right
after the current one was created.  We then had to go back to the
terminal, find the right folder, and run ember destroy with the right
parameters.  That's a tad of a mess.  The new implementation has support
for destroying elements too.

The code could benefit from being made a tad more generic though.  It
currently has quite som duplication.