Remove --pod default option due to confusion
If .ember-cli has usePods set to true, suppling the --pod option will
negate the set default.  Thus, if .ember-cli has usePods set to true,
and you call generate model with the --pod option, it will generate a
non-pod model.  As this is rather confusing, the --pod is removed as a
default option when pods are enabled.