Add support for importing statements
The support is attached to `C-c C-c i u' and makes it easy to migrate
from old Ember.something definitions to the new import statements.
The content is based on
https://github.com/ember-cli/ember-rfc176-data#contents

Ideally, this would be split in multiple packages so they can be
required separately.  This may be handled in a separate project.