Push xref marker upon jumping when it exists
If xref-push-marker-stack is an available function, we now call that
function when jumping between known files in ember-mode.  This
provides more seamless integration with the lsp functionality.

This is mainly a PoC.  Effective implementation may change upon
further reading.

There is still a wee bit of doubt as to whether we should mark the
location when jumping to, rather than when jumping from.  Further
reading and experimentation may make us implement this feature
slightly differently.  Feel free to create an issue if the behaviour
feels wrong.